test: CopyToFiles unit and IT tests refactor: CopyToFiles class separate methods and construct logic in handle chore: minor file formatting
